new updates

+++ Algorithm4/include/searching_and_sorting.h
@@ -4,5 +4,5 @@
 
 void merge_karo(double *, int, int, int);
 void sort_karo(double *, int, int, int);
-int binary_search_thr(double *thr_arr, int key, int i, int n);
+int binary_search_thr(double *thr_arr, double key, int i, int n);
 #endif
